vision/requirements.txt
paho-mqtt
cv2
